And then PR was a zombie

+++ code/modules/turbolift/turbolift.dm
@@ -153,15 +153,17 @@
 	if(!istype(origin) || !istype(destination) || (origin == destination))
 		return 0
 
-	for(var/turf/T in destination)
-		for(var/atom/movable/AM in T)
-			if(istype(AM, /mob/living))
-				var/mob/living/M = AM
-				M.gib()
-			else if(istype(AM, /mob/zshadow))
-				AM.Destroy()		//prevent deleting shadow without deleting shadow's shadows
-			else if(AM.simulated && !(istype(AM, /mob/observer)))
-				qdel(AM)
+	if(!moving_upwards)
+		for(var/turf/T in destination)
+			set waitfor = FALSE		// no check ticks until we get better shuttlecode.
+			for(var/atom/movable/AM in T)
+				if(istype(AM, /mob/living))
+					var/mob/living/M = AM
+					M.gib()
+				else if(istype(AM, /mob/zshadow))
+					AM.Destroy()		//prevent deleting shadow without deleting shadow's shadows
+				else if(AM.simulated && !(istype(AM, /mob/observer)))
+					qdel(AM)
 
 	origin.move_contents_to(destination)
